factor out API calls into /lib
